add py3 version to zipimport
