--- stdlib/2.7/zipimport.pyi
@@ -1,25 +1,0 @@
-"""Stub file for the 'zipimport' module."""
-
-from typing import Dict, Optional
-from types import CodeType, ModuleType
-
-class ZipImportError(ImportError):
-    pass
-
-_zip_directory_cache = ...  # type: Dict[str, dict]
-
-class zipimporter(object):
-    archive = ...  # type: str
-    prefix = ...  # type: str
-    _files = ...  # type: Dict[str, tuple]
-    def __init__(self, path: str) -> None:
-        raise ZipImportError
-    def find_module(self, fullname: str, path: str = ...) -> Optional['zipimporter']: ...
-    def get_code(self, fullname: str) -> CodeType: ...
-    def get_data(self, fullname: str) -> str:
-        raise IOError
-    def get_filename(self, fullname: str) -> str: ...
-    def get_source(self, fullname: str) -> str: ...
-    def is_package(self, fullname: str) -> bool: ...
-    def load_module(self, fullname: str) -> ModuleType: ...
-
